blockly_server/requirements.txt
Jinja2
flask
pyyaml
SQLAlchemy
Flask-SQLAlchemy
SQLAlchemy-serializer
Flask-Cors
eventlet
Flask_SocketIO
Flask-Babel
simple-websocket
requests
websocket-client